glpMsgTbl stores 64-bit pointers to turn_t variables. This won't fit inside int.

+++ Source/nthread.cpp
@@ -9,7 +9,7 @@ static CCritSect sgMemCrit;
 DWORD gdwDeltaBytesSec;
 BOOLEAN nthread_should_run;
 DWORD gdwTurnsInTransit;
-int glpMsgTbl[MAX_PLRS];
+uintptr_t glpMsgTbl[MAX_PLRS];
 unsigned int glpNThreadId;
 char sgbSyncCountdown;
 int turn_upper_bit;
